Ensure that _align_to_power_of_2 DTRT on ppc64.
git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@1775 8bdb8f1d-59ce-db11-9e41-0016172a54ae

+++ lisp-kernel/lisp.h
@@ -29,7 +29,7 @@
 static inline unsigned long
 _align_to_power_of_2(unsigned long n, unsigned power)
 {
-  unsigned align = (1<<power) -1;
+  unsigned long align = (1<<power) -1;
 
   return (n+align) & ~align;
 }
